feat: update all code

+++ distances/distance_type.py
@@ -12,7 +12,7 @@ class DistanceType(Enum):
 
 def calculate_distance(point_1, point_2, distance_type=None):
     if distance_type == DistanceType.EUCLIDEAN:
-        return float(math.sqrt((point_2[0] - point_1[0]) ** 2 + (point_2[1] - point_1[1]) ** 2) + 0.5)
+        return int(math.sqrt((point_2[0] - point_1[0]) ** 2 + (point_2[1] - point_1[1]) ** 2) + 0.5)
     if distance_type == DistanceType.MANHATTAN or not distance_type:
         return abs(point_1[0] - point_2[0]) + abs(point_1[1] - point_2[1])
     if distance_type == DistanceType.HAVERSINE:
